playbooks: refactor playbooks to remove unnecessary 'become: true' statements
When you set `become: true` at the playbook level in an Ansible playbook, this setting applies to all tasks across all roles within that playbook, running them with sudo privileges. This might not be desirable for certain tasks or roles, such as those involving kubectl or other operations that should not run as the superuser.

Signed-off-by: Abhishek Gaikwad <gaikwadabhishek1997@gmail.com>