playbooks: add playbook for node replacement
This playbook automates the process of replacing a node in the Kubernetes cluster. It handles unlabeling the old node, deleting the target and proxy pods, PVCs, and PVs, and labeling the new node for proper pod deployment. Additionally, it integrates the  role to ensure new PVs and PVCs are created for the target node.

Signed-off-by: Abhishek Gaikwad <gaikwadabhishek1997@gmail.com>